+++ src/component/Menu/index.js
@@ -1,13 +1,4 @@
 import Menu from './Menu';
-// import Divider from './Divider';
-// import ItemGroup from './ItemGroup';
-// import MenuItem from './MenuItem';
-// import SubMenu from './SubMenu';
 import './_menu.scss';
-//
-// Menu.Divider = Divider;
-// Menu.ItemGroup = ItemGroup;
-// Menu.MenuItem = MenuItem;
-// Menu.SubMenu = SubMenu;
 
 export default Menu;